Refatoração de Layout e configuração do BindingAdapter

+++ app/src/main/java/com/allisson/appabracao/ui/listaongs/ListaOngsViewModel.kt
@@ -17,7 +17,7 @@ class ListaOngsViewModel : ViewModel() {
                     ong.id,
                     ong.data["nome"].toString(),
                     ong.data["endereco"].toString(),
-                    ong.data["qtdPessoas"].toString().toInt(),
+                    ong.data["qtdPessoas"].toString(),
                     ong.data["foto"].toString(),
                     ong.data["telefone"].toString(),
                     ong.data["email"].toString(),
